+++ Sources/Tabman/TabmanBar/Configuration/TabmanBarConfigHandler.swift
@@ -1,0 +1,43 @@
+//
+//  TabmanBarConfigHandler.swift
+//  Tabman
+//
+//  Created by Merrick Sapsford on 21/11/2017.
+//  Copyright © 2017 UI At Six. All rights reserved.
+//
+
+import Foundation
+
+/// Update handler protocol for TabmanBarConfig updates.
+internal protocol TabmanBarConfigHandler: class {
+    
+    /// The config had its style updated.
+    ///
+    /// - Parameters:
+    ///   - config: The config.
+    ///   - style: The new style.
+    func config(_ config: TabmanBar.Config, didUpdate style: TabmanBar.Style)
+    
+    /// The config had its location updated.
+    ///
+    /// - Parameters:
+    ///   - config: The config.
+    ///   - location: The new location.
+    func config(_ config: TabmanBar.Config, didUpdate location: TabmanBar.Location)
+    
+    /// The config had its items updated.
+    ///
+    /// - Parameters:
+    ///   - config: The config.
+    ///   - items: The new items.
+    func config(_ config: TabmanBar.Config, didUpdate items: [TabmanBar.Item]?)
+    
+    /// The config had its appearance config updated.
+    ///
+    /// - Parameters:
+    ///   - config: The config.
+    ///   - appearance: The new appearance config.
+    func config(_ config: TabmanBar.Config, didUpdate appearance: TabmanBar.Appearance)
+    
+    func config(_ config: TabmanBar.Config, didUpdate behaviors: [TabmanBar.Behavior]?)
+}
